Update Terakhir - 01/07/2019

+++ website/rusunawa/admin/index_sambutan.php
@@ -67,30 +67,7 @@
 
     <div id="wrapper">
 
-      <!-- Sidebar -->
-      <ul class="sidebar navbar-nav">
-        <li class="nav-item active">
-          <a class="nav-link" href="index.php">
-            <i class="fas fa-fw fa-tachometer-alt"></i>
-            <span>Dashboard</span>
-          </a>
-        </li>
-        <li class="nav-item dropdown">
-          <a class="nav-link dropdown-toggle" href="#" id="pagesDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-            <i class="fas fa-fw fa-folder"></i>
-            <span>Data Master</span>
-          </a>
-          <div class="dropdown-menu" aria-labelledby="pagesDropdown">
-            <!-- <a class="dropdown-item" href="#">User</a> -->
-            <a class="dropdown-item" href="index_tower_a.php">Tower A</a>
-            <a class="dropdown-item" href="index_tower_b.php">Tower B</a>
-            <a class="dropdown-item" href="index_tower_c.php">Tower C</a>
-            <a class="dropdown-item" href="index_booking.php">Booking</a>
-            <a class="dropdown-item" href="index_berita.php">Update Berita</a>
-            <a class="dropdown-item" href="index_sambutan.php">Update Sambutan</a>
-            <a class="dropdown-item" href="index_galery.php">Update Gallery</a>
-        </li>
-      </ul>
+      <?php include('sidebar.php'); ?>
 
       <div id="content-wrapper">
 
@@ -104,73 +81,7 @@
             <li class="breadcrumb-item active">Selamat Datang Di Rusunawa Depok</li>
           </ol>
 
-          <!-- Icon Cards-->
-          <div class="row">
-            <div class="col-xl-3 col-sm-6 mb-3">
-              <div class="card text-white bg-primary o-hidden h-100">
-                <div class="card-body">
-                  <div class="card-body-icon">
-                    <i class="fas fa-fw fa-comments"></i>
-                  </div>
-                  <div class="mr-5">Total Waiting</div>
-                </div>
-                <a class="card-footer text-white clearfix small z-1" href="#">
-                  <span class="float-left">View Details</span>
-                  <span class="float-right">
-                    <i class="fas fa-angle-right"></i>
-                  </span>
-                </a>
-              </div>
-            </div>
-            <div class="col-xl-3 col-sm-6 mb-3">
-              <div class="card text-white bg-warning o-hidden h-100">
-                <div class="card-body">
-                  <div class="card-body-icon">
-                    <i class="fas fa-fw fa-list"></i>
-                  </div>
-                  <div class="mr-5">Booking</div>
-                </div>
-                <a class="card-footer text-white clearfix small z-1" href="#">
-                  <span class="float-left">View Details</span>
-                  <span class="float-right">
-                    <i class="fas fa-angle-right"></i>
-                  </span>
-                </a>
-              </div>
-            </div>
-            <div class="col-xl-3 col-sm-6 mb-3">
-              <div class="card text-white bg-success o-hidden h-100">
-                <div class="card-body">
-                  <div class="card-body-icon">
-                    <i class="fas fa-fw fa-shopping-cart"></i>
-                  </div>
-                  <div class="mr-5">Update Berita</div>
-                </div>
-                <a class="card-footer text-white clearfix small z-1" href="#">
-                  <span class="float-left">View Details</span>
-                  <span class="float-right">
-                    <i class="fas fa-angle-right"></i>
-                  </span>
-                </a>
-              </div>
-            </div>
-            <div class="col-xl-3 col-sm-6 mb-3">
-              <div class="card text-white bg-danger o-hidden h-100">
-                <div class="card-body">
-                  <div class="card-body-icon">
-                    <i class="fas fa-fw fa-life-ring"></i>
-                  </div>
-                  <div class="mr-5">Update Galery</div>
-                </div>
-                <a class="card-footer text-white clearfix small z-1" href="#">
-                  <span class="float-left">View Details</span>
-                  <span class="float-right">
-                    <i class="fas fa-angle-right"></i>
-                  </span>
-                </a>
-              </div>
-            </div>
-          </div>
+          <?php include('nav-button.php'); ?>
 
 
               <div id="content-wrapper">
@@ -209,29 +120,28 @@
                       <th>Action</th>
                     </tr>
                   </thead>
-<?php
-include 'connection.php';
-$query="SELECT*FROM update_sambutan";
-$result=mysqli_query($conn,$query);
-while ($data=mysqli_fetch_array($result)) {
-echo "<tr align='center'>
-<td>".$data['kode_sambutan']."</td>
-<td><img src='images/".$data['gambar']."' width='100' height='100'></td>
-<td>".$data['kata_sambutan']."</td>"; ?>
-<td>
-    <li name="delete" class="btn btn btn-sm" value="delete" title="Hapus Data">
-    <a href="hapus_sambutan.php?id=<?php echo $data['kode_sambutan'];?>">
-         <i class="fas fa-trash"></i>
-    </li>
-    <li name="update" class="btn btn btn-sm" value="update" title="Edit Data">
-    <a href="form_update_sambutan.php?id=<?php echo $data['kode_sambutan'];?>">
-          <i class="fas fa-paperclip"></i>
-    </li>
-</td>
-<?php
-}
-echo "</table>";
-?>
+                  <?php
+                  include 'connection.php';
+                  $query="SELECT*FROM update_sambutan";
+                  $result=mysqli_query($conn,$query);
+                  while ($data=mysqli_fetch_array($result)) {
+                  echo "<tr align='center'>
+                  <td>".$data['kode_sambutan']."</td>
+                  <td><img src='images/".$data['gambar']."' width='100' height='100'></td>
+                  <td>".$data['kata_sambutan']."</td>"; ?>
+                  <td>
+                      <li name="delete" class="btn btn btn-sm" value="delete" title="Hapus Data">
+                      <a href="hapus_sambutan.php?id=<?php echo $data['kode_sambutan'];?>" onClick="return confirm('Anda Yakin Ingin Menghapus Data <?php echo $data['kode_sambutan'];?> ?')">
+                           <i class="fas fa-trash"></i>
+                      </li>
+                      <li name="update" class="btn btn btn-sm" value="update" title="Edit Data">
+                      <a href="form_update_sambutan.php?id=<?php echo $data['kode_sambutan'];?>">
+                            <i class="fas fa-paperclip"></i>
+                      </li>
+                  </td>
+                  <?php
+                    }
+                  ?>
                 </table>
               </div>
